将实验室检查，免疫组化，分子检测提出作为公用模块，增加进治疗信息中
